test: move test config fixtures so they can be accessed by search API tests #258
